requirements.txt
click
python-dotenv
binance-futures-connector